Add files via upload
Processes information from a csv file about Pokemon and their statistics and organizes the information into a 2-level dictionary. This poke_dict is used to compute maximum averages for a given statistic like speed, hp, etc. A user can enter a stat and this program with return the type of Pokemon with that max average for that stat and then the numerical value.