Add files via upload
This program takes a file of women's NCAA D1 basketball programs with
data for the name, conference, wins, and losses. It parses the file using the
Team class to create a team object. It adds teams to the ConferenceSet class, 
a collection of teams in one conference, and calculates the conference with
the highest win ratio.